feat: organiza arquitetura do software
