+++ models/atividades.py
@@ -1,0 +1,11 @@
+from sqlalchemy import Column, Integer, String, ForeignKey
+from sqlalchemy.orm import relationship
+from models.database import Base
+
+
+class Atividades(Base):
+    __tablename__ = 'atividades'
+    id = Column(Integer, primary_key=True)
+    nome = Column(String(80))
+    pessoa_id = Column(Integer, ForeignKey('pessoas.id'))
+    pessoa = relationship("Pessoas")
