+++ models/database/__init__.py
@@ -1,0 +1,19 @@
+from sqlalchemy.orm import scoped_session, sessionmaker
+from sqlalchemy.ext.declarative import declarative_base
+from sqlalchemy import create_engine
+import sqlalchemy.orm
+
+engine = create_engine('sqlite:///atividades.db')
+db_session = scoped_session(sessionmaker(autocommit=False,
+                                         bind=engine))
+
+Base = sqlalchemy.orm.declarative_base()
+Base.query = db_session.query_property()
+
+
+def init_db():
+    Base.metadata.create_all(bind=engine)
+
+
+if __name__ == '__main__':
+    init_db()
