+++ models/usuarios.py
@@ -1,0 +1,9 @@
+from sqlalchemy import Column, Integer, String
+from models.database import Base
+
+
+class Usuarios(Base):
+    __tablename__ = 'usuarios'
+    id = Column(Integer, primary_key=True)
+    login = Column(String(20), unique=True)
+    senha = Column(String(150))
