+++ services/utils.py
@@ -1,0 +1,7 @@
+import bcrypt
+
+def criptografar_senha(senha: str) -> bytes:
+    salt = bcrypt.gensalt()
+    password_encoded = senha.encode('utf-8')
+    hash = bcrypt.hashpw(password=password_encoded, salt=salt)
+    return hash
